Add Docker

+++ ToDoTask/Controllers/HomeController.cs
@@ -29,7 +29,7 @@ namespace ToDoTask.Controllers
         public IActionResult TaskDetails(int id)
         {
             var task = _taskService.GetTaskById(id);
-            if (task == null)
+            if (task == default)
             {
                 return NotFound();
             }
@@ -37,9 +37,18 @@ namespace ToDoTask.Controllers
             return View(task);
         }
 
+        [HttpPost]
         public IActionResult AddTask(TaskModel task)
         {
-            _taskService.AddTask(task);
+            _logger.LogInformation($"AddTask called with title: {task.Title}");
+            if (task.Title == default) 
+            {
+                _logger.LogInformation($"AddTask denied, title is null: {task.Title}");
+            }
+            else
+            {
+                _taskService.AddTask(task);
+            }           
             return RedirectToAction("Index");
         }
 
@@ -48,22 +57,35 @@ namespace ToDoTask.Controllers
         {
             _logger.LogInformation($"EditTask called with title: {model.Title}, newTitle: {model.NewTitle}, isDone: {model.IsCompleted}");
 
+            if (model.Title == default || model.NewTitle == default)
+            {
+                _logger.LogInformation("Invalid request. Title cannot be null.");
+                return Json(new { success = false, errorMessage = "Title cannot be null." });
+            }
+
             var editedTask = _taskService.EditTask(model.Title, model.NewTitle, model.IsCompleted);
 
-            if (editedTask != null)
+            if (editedTask != default)
             {
                 _logger.LogInformation($"Task edited successfully. Title: {editedTask.Title}, IsDone: {editedTask.IsCompleted}");
                 return Json(new { success = true, editedTask });
             }
 
             _logger.LogInformation("Task not found for editing.");
-            return Json(new { success = false });
+            return Json(new { success = false, errorMessage = "Task not found for editing." });
         }
+
 
         [HttpPost]
         public IActionResult DeleteTask(TaskModel model)
         {
             _logger.LogInformation($"DeleteTask called with title: {model.Title}");
+
+            if (model.Title == default)
+            {
+                _logger.LogInformation("Invalid request. Title cannot be null.");
+                return BadRequest(new { success = false, errorMessage = "Title cannot be null." });
+            }
 
             var success = _taskService.DeleteTask(model.Title);
 
@@ -74,8 +96,9 @@ namespace ToDoTask.Controllers
             }
 
             _logger.LogInformation($"Failed to delete task. Task \"{model.Title}\" not found.");
-            return Json(new { success = false });
+            return Json(new { success = false, errorMessage = $"Task \"{model.Title}\" not found." });
         }
+
 
 
         public IActionResult Editing()
